pathfinder-ai/backend/requirements.txt
# Pathfinder AI — Backend dependencies
# Run from D:\browser_agent\pathfinder-ai\:
#   pip install -r backend/requirements.txt

# ── FastAPI ───────────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.18

# ── LangChain + Groq ──────────────────────────────────────────────────────────
langchain>=0.2.0,<0.3.0
langchain-groq>=0.1.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-classic>=0.0.1

# ── Search ────────────────────────────────────────────────────────────────────
serpapi>=0.1.5
google-search-results>=2.4.0

# ── File Processing ───────────────────────────────────────────────────────────
PyPDF2>=3.0.0
python-docx>=1.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
